Compact InputSection from 112 to 96 bytes
InputSection is the highest-count struct in a link - a debug clang link
has ~4.4M of them - and it is walked in dead-strip, ICF, relocation
scanning, output-section assignment, thunk layout and the copy phase,
so its width drives the cache traffic of all of them. mold-rust keeps
its InputSection at 64 bytes for exactly this reason; ours had grown to
112 with three oversized fields:

  - obj and osec were usize (8 bytes each). They index the object and
    chunk arenas, neither of which approaches 2^32, so they become u32
    (their usize::MAX sentinels become u32::MAX). mold-rust indexes
    with u32 too.
  - replacement was Option<usize> (16 bytes) for a literal-merge
    redirect. A u32 subsection index with a NO_REPLACEMENT (u32::MAX)
    sentinel is 4 bytes and needs no niche.

That is 20 bytes off the fields, 16 off the struct after alignment
(112 -> 96), asserted at compile time. Call sites cast at the arena
boundary (obj/osec as usize) and test the sentinel instead of Option.
322 integration tests pass and the linked binaries run.

+++ src/arch/arm64.rs
@@ -157,7 +157,7 @@ impl Arch for Arm64 {
                     if !isec.is_alive || isec.output_offset == u64::MAX {
                         continue 'hint;
                     }
-                    let chunk = &ctx.chunks[isec.osec];
+                    let chunk = &ctx.chunks[isec.osec as usize];
                     locs.push((
                         (chunk.hdr.fileoff + isec.output_offset + off) as usize,
                         chunk.hdr.addr + isec.output_offset + off,
@@ -406,7 +406,7 @@ impl Arch for Arm64 {
         base: u64,
         buf: &mut [u8],
     ) {
-        let obj = ctx.isecs[isec_id].obj;
+        let obj = ctx.isecs[isec_id].obj as usize;
         let mut i = 0;
         while i < rels.len() {
             let r = &rels[i];
@@ -459,7 +459,7 @@ impl Arch for Arm64 {
                         } else {
                             let isec = &ctx.isecs[isec_id];
                             let thunk_addr =
-                                ctx.chunks[isec.osec].hdr.addr + r.thunk_off as u64;
+                                ctx.chunks[isec.osec as usize].hdr.addr + r.thunk_off as u64;
                             val = thunk_addr.wrapping_sub(p) as i64;
                         }
                     }
